Don't throw response exception if unable to resolve authorized user
That should be handled by shiro using the @RequiresAuthentication annotation, which is implemented using a method interceptor, meaning it happens after argument resolution. Now, if an endpoint does not require authentication but has a required `@AuthorizedUser User` then an NPE is thrown. This would be a programmer error, but security doesn't suffer from it.